backend/car_parking/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
sqlparse==0.4.3
Django==3.2.3
djangorestframework==3.12.4
